+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## 0.1.7 [unreleased]
+## 0.1.8 [unreleased]
+
+### Changed
+
+### Added
+
+## 0.1.7 [15/06/2019]
 
 ### Changed
 - Latex_labels: Moved the latex labels into pesummary.core.plots.latex_labels
@@ -27,6 +33,9 @@
   PESummary meta file.
 - Added a `--no_ligo_skymap` option which will prevent ligo.skymap from
   generating a skymap
+- Added a `--gwdata` flag which allows the user to pass a gw strain cache file.
+  If a valid cache file is passed, the timeseries is plotted and the maxL
+  waveform superimposed.
 
 ## 0.1.6 [13/05/2019]